Change a package name
