--- app/src/androidTest/java/com/example/android/architecture/blueprints/todoapp/util/SingleExecutors.java
@@ -1,39 +1,0 @@
-/*
- * Copyright (C) 2017 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.android.architecture.blueprints.todoapp.util;
-
-import android.support.annotation.NonNull;
-
-import com.amatda.util.AppExecutors;
-
-import java.util.concurrent.Executor;
-
-/**
- * Allow instant execution of tasks.
- */
-public class SingleExecutors extends AppExecutors {
-    private static Executor instant = new Executor() {
-        @Override
-        public void execute(@NonNull Runnable command) {
-            command.run();
-        }
-    };
-
-    public SingleExecutors() {
-        super(instant, instant, instant);
-    }
-}
