+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.android.architecture.blueprints.todoapp">
+    package="com.amatda">
 
     <application
         android:allowBackup="false"
