+++ app/src/main/java/com/amatda/tasks/TasksActivity.java
@@ -28,8 +28,8 @@ import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
 
-import com.example.android.architecture.blueprints.todoapp.Injection;
-import com.example.android.architecture.blueprints.todoapp.R;
+import com.amatda.Injection;
+import com.amatda.R;
 import com.amatda.statistics.StatisticsActivity;
 import com.amatda.util.ActivityUtils;
 import com.amatda.util.EspressoIdlingResource;
